+++ lib/rodauth/rails.rb
@@ -15,11 +15,7 @@ module Rodauth
 
     class << self
       def rodauth(name = nil, query: nil, form: nil, account: nil, **options)
-        auth_class = app.rodauth(name)
-
-        unless auth_class
-          fail ArgumentError, "undefined rodauth configuration: #{name.inspect}"
-        end
+        auth_class = app.rodauth!(name)
 
         LOCK.synchronize do
           unless auth_class.features.include?(:internal_request)